Restructured file structure into respectable packages.
